requirements.txt
pandas>=1.5
numpy>=1.23
scikit-learn>=1.2
xgboost>=1.7
lightgbm>=3.3
matplotlib>=3.6
seaborn>=0.12
tqdm>=4.64
PyYAML>=6.0
scipy>=1.9
# Hyperfast (optional). If unavailable, the wrapper falls back automatically.
# hyperfast
